chore(dashboard): rename + move components, and lift store up into testing-ground so it is accessible throughout dashboard
