+++ packages/dashboard/src/components/iot-dashboard-grid/getSelectionBox.spec.ts
