+++ packages/dashboard/src/components/iot-dashboard-grid/getSelectionBox.ts
