+++ packages/dashboard/src/components/iot-dashboard-grid/iot-dashboard-grid.css
@@ -1,0 +1,33 @@
+@import "../variables.css";
+
+:host {
+  display: block;
+}
+
+.container {
+  position: absolute;
+  height: 100%;
+  margin: 0.5em;
+
+  /**  establish a depth context */
+  z-index: 1;
+}
+
+.select-rect {
+  border: var(--selection-border-width) solid var(--selection-color);
+  position: absolute;
+}
+
+.grid-image {
+  z-index: -999999999;
+  position: absolute;
+
+  /** Offset to make the dot fall in the top left corner */
+  right: 4px;
+  top: -4px;
+  background-position: top left;
+  background-image: url("grid-dot.svg");
+  background-repeat: repeat;
+  width: 100%;
+  height: 100%;
+}
