+++ packages/dashboard/src/components/iot-dashboard-grid/iot-dashboard-widget/iot-dashboard-widget.css
