+++ packages/dashboard/src/components/iot-dashboard-grid/iot-dashboard-widget/iot-selection-box-anchor.tsx
