+++ packages/dashboard/src/components/iot-dashboard-grid/iot-selection-box/iot-selection-box.css
