+++ packages/dashboard/src/components/iot-dashboard-grid/trimWidgetPosition.spec.ts
