+++ packages/dashboard/src/components/iot-dashboard-grid/trimWidgetPosition.ts
