+++ packages/dashboard/src/components/iot-dashboard/iot-dashboard.css
@@ -1,33 +1,15 @@
-@import "../variables.css";
-
-:host {
-  display: block;
+.iot-dashboard label {
+  margin-right: 0.5em;
+  display: inline-block;
 }
 
-.container {
-  position: absolute;
-  height: 100%;
-  margin: 0.5em;
-
-  /**  establish a depth context */
-  z-index: 1;
-}
-
-.select-rect {
-  border: var(--selection-border-width) solid var(--selection-color);
-  position: absolute;
-}
-
-.grid-image {
-  z-index: -999999999;
-  position: absolute;
-
-  /** Offset to make the dot fall in the top left corner */
-  right: 4px;
-  top: -4px;
-  background-position: top left;
-  background-image: url("grid-dot.svg");
-  background-repeat: repeat;
-  width: 100%;
-  height: 100%;
+.iot-dashboard-dummy-content {
+  width: calc(100% - 40px);
+  height: calc(100% - 40px);
+  min-height: calc(100vh - 210px);
+  background-color: rgba(0 0 0 / 1%);
+  padding: 20px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
 }
