--- packages/dashboard/src/testing/testing-ground/testing-ground.css
@@ -1,15 +1,0 @@
-label {
-  margin-right: 0.5em;
-  display: inline-block;
-}
-
-.dummy-content {
-  width: calc(100% - 40px);
-  height: calc(100% - 40px);
-  min-height: calc(100vh - 210px);
-  background-color: rgba(0 0 0 / 1%);
-  padding: 20px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
